Add ScoresIterator class to avoid code duplication
Instead of having two private classes inside the player board states,
that differed only for the scores set and assignment logic,
we use a configurable class that provides an iterator according to the
provided scores set, starting from a specified index.
In this way we can shift the scores set with respect to the player's deaths
in regular mode, and always assign the same scores in frenzy mode.