Implement methods to reload, unload and reset weapons
Also added a private method to the player class to convert a list
of AmmoCubeCost objects into a map of AmmoCube objects:
in case the cost contains gray cubes, the method compensates their amount
with cubes taken from the player's ammo box and throws an exception if
there is not sufficient ammo for the payment.
Added an overloaded payAmmo method that accepts a list of AmmoCubeCost objects
and automatically performs conversion.
Added method to AmmoCubeCost that returns the corresponding equal AmmoCube value.