Implement turn management and requests cancelling
Added a cancel method to the choice handler interface, to handle failures in retrieving the choice.
Implented methods to select the next turn and set it as the current one.
Also added logic in next turn selection to trigger game ending.
Added loop over player boards to switch them to frenzy state when frenzy mode is triggered