Add adjacencies rebuilding when inserting a square
Fixed method to insert a square in the board in order not to break the adjacency map of its neghbours (if present): when inserting a new square that replaces another one we check for each cardinal direction and make the corresponding neighbours point to the new square.
Added helper method to the CardinalDirection class to retrieve the opposite cardinal direction of the current one.