Implement method to retrieve scoped squares in Board
Added FunctionalInterface to represent a lambda taking three arguments.
Added the aforementioned functional interface to the visibility enumerator, to make it filter a set of squares given the starting square, the current board and its own  value.
Implemented method to retrieve scoped squares according to visibility, cardinal direction and distance.