Add method to retrieve squares reachable by a square in a certain number of steps
Instead of computing the set while determing whether two squares are connected, and in order to make the result reusable in other operations, the isReachable method hs been split up in another method called getNavigableSquares, that provides the set of squares that can be reached from the starting one in at most the provided number of steps.
To then check whether a square is reachable by another one we simply get this method's result and check if it's contained in it.