Refactor ChoiceHandler to accept the related player and the controller
Added player username to the parameters of the handle method, to determine who performed the choice and apply its effects on him.
Added also the controller among the said parameters in order to access the match and apply possible side-effects.
Also changed type of choices collection in Request to Collection, to allow specialized types (e.g. List and Set) when creating the request.