Start implementing handlers for grab and movement events
Implemented handlers for GrabEvent and MoveEvent (to be finished).
Added specialized requests for:
* destination square
* grabbed weapon card
Added specialized choice handlers to handle choices of the aforementioned requests.
Changed signature of the handle method to accept a Controller instead of a Match object, in order to be able to set a new pending request.
Added static attribute in ControllerImpl describing the maximum move steps allowed for a player (used to determine the set of destination squares the user can choose among).