Add lobby screen and spinner animation in CLIHelper
Added lobby page to wait for the match to start.
Added spinner animation in CLI mode.
Refactored print method not to print a newline character and added a println method instead.
Adedd methods to UIRenderer to manage lobby and to prepare the game screen.
Added scenes to the ViewScene enumerator.