Add char representation to connection type and add method to render squares in CLI
Added the corresponding character to the enumerator of connection types.
Refactored adjacency map in SquareView to just use the direction and connection type.
Added method that renders a square in a string format (List<List<String>>),
that allows it to be printed in a terminal and to be placed side by side.