Add methods to render weapon cards in CLI and to optimize width of strings and lists
Added methods that, given a string or a list, split them into chunks that fit the provided width, in order to make text and objects' lists reflow automatically in rendered content.
Added method to generate a textual rendering of a weapon card (effects have to be added yet).
Added constant attributes for borders and corners, along with an escape sequence to make text bold.
Added methods and attributes to retrieve the ANSI and HEX codes for the color of ammo cubes.