Add class to handle selection events from selectable overlays in GUI
Added class extending a consumer of SelectableOverlay, in order to handle when a certain element is selected and reply to a choice request accordingly.
Added choice requests management in ClientView, to put pending requests in a queue and to expose a method to be called when a request has been completely handled.
Also fixed model updates not being redirected to ModelManager.