Add method to print the list of all the players with their color and symbol in CLI
Added a method that renders a table showing the list of all the players participating in the match, along with their color and symbol, to serve as a reference during the match and to avoid printing whole names across the screen.
Added also an helper method in CLIHelper to truncate strings to make them fit a provided width.