Implement cancellation of selection in GUI and handle operation requests
When setting a new selection handler in the client view, if an handler is already present it is cancelled before being replaced by the new one.
Implemented method to cancel the current selection, by setting a null handler to the client view and fixed potential bug when doing so (setting the view on a null handler).
Implemented then the method to ask the user to select an operation to perform, acting to the operations track control present in the game screen.