Add class to refresh rendered elements according to received update
Added a class implementing the ModelUpdateVisitor interface to refresh the rendered game elements according to the received update, lifting the ModelManager from this burden and removing its dependency from the ClientView.
Added methods in the UIRenderer interface to refresh specific elements of the game, in order to provide a granular refresh technique (instead of just repainting the whole screen).