Add methods to notify user about disconnection and reconnect
Added new scene to represent a disconnected state.
Added method in renderer to show a screen telling the user about its disconnection, which lets him/her reconnect to the game.
Since when a player logs in it receives the initial update, after logging back in we just wait for it to be received, which triggers the rendering of the match screen.
Implemented then a scene in GUI to perform reconnection.