Add weapons info pane in game screen and fix weapon updates handling
Added custom control to show information about the weapons owned by the players,
that also allows the user to expand the details of a weapon card, showing its graphical representation in a new window.
Added a controller to manage windows used to show images.
Added method in UIRenderer to refresh the rendered information about owned weapons.