Add alert when a player is suspended and fix updates mechanism
Added method in UIRenderer to show a preemption screen, used to notify the user that he/she is disconnected or he/she has been suspended from the match, after which the client is closed.
Fixed view generation of Player, where score and state were missing.
Removed useless updates broadcasting (e.g. when they were already sent as a side effect of a previously called method and did not differ in terms of information).