Fix marks not being added to damage when shooting
When performing damage on a player, marks are now taken into account and
added to the damage amount, resetting then the count of marks from the
shooter.
Also added method to reset the marks from a given player (w/ tests).
Also fixed wrong computed count of marks when generating the view of a
playerboard.
Also fixed missing marks on the first target of "hellion" weapon.